requirements.txt
tensorflow
pandas
numpy
ultralytics
opencv-python
cvzone
streamlit_webrtc
streamlit
pillow
watch dog
